refactor: :truck: rename the `create_readme_text` module to `build_`
